# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Modular API server. Run in platform or tenant mode via -mode flag.",
        "title": "Go Web API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/users": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get paginated list of users with optional search and filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "List users with pagination and filters",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "description": "Items per page (default: 20, max: 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search in email, first name, last name",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "active",
                            "inactive",
                            "expired"
                        ],
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "created_at",
                            "email",
                            "first_name",
                            "last_name"
                        ],
                        "type": "string",
                        "description": "Sort field (default: created_at)",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "description": "Sort order (default: desc)",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_users_models.UserListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_core_errors.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_core_errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a single user by their ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get user by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_core_models.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_core_errors.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_core_errors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_core_errors.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Soft delete a user (sets deleted_at timestamp and invalidates sessions). Respects ALLOW_USER_DELETION and ENABLE_SOFT_DELETE config.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Soft delete user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Deletion details",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_users_models.SoftDeleteUserDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_core_errors.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_core_errors.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "User deletion is disabled",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_core_errors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_core_errors.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "josex_web_modules_core_errors.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Error code\nexample: 400",
                    "type": "string"
                },
                "detail": {
                    "description": "Error details\nexample: {\"email\": \"Invalid email format\"}"
                },
                "error": {
                    "description": "Error message\nexample: Bad Request"
                },
                "message": {
                    "description": "Translated error message"
                }
            }
        },
        "josex_web_modules_core_models.User": {
            "type": "object",
            "required": [
                "email",
                "first_name",
                "id",
                "last_name"
            ],
            "properties": {
                "bio": {
                    "type": "string"
                },
                "birthday": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "expiration_date": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "last_name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "profile_picture_url": {
                    "type": "string"
                },
                "tenant_role": {
                    "type": "string"
                },
                "website_url": {
                    "type": "string"
                }
            }
        },
...
        "josex_web_modules_users_models.SoftDeleteUserDto": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "reason": {
                    "description": "Optional deletion reason",
                    "type": "string",
                    "maxLength": 500
                }
            }
        },
        "josex_web_modules_users_models.UserListResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "description": "Items per page",
                    "type": "integer"
                },
                "page": {
                    "description": "Current page",
                    "type": "integer"
                },
                "total": {
                    "description": "Total count of users",
                    "type": "integer"
                },
                "total_pages": {
                    "description": "Total pages",
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/josex_web_modules_core_models.User"
                    }
                }
            }
        }
    }
}